[no-unnecessary-condition]: add checkTruthinessAssertions
