+++ packages/eslint-plugin/docs/rules/no-unnecessary-condition.mdx
@@ -90,6 +90,28 @@ for (; true; ) {}
 do {} while (true);
 ```
 
+### `checkTruthinessAssertions`
+
+Example of additional incorrect code with `{ checkTruthinessAssertions: true }`:
+
+```ts option='{ "checkTruthinessAssertions": true }' showPlaygroundButton
+function assert(condition: unknown): asserts condition {
+  if (!condition) {
+    throw new Error('Condition is falsy');
+  }
+}
+
+assert(false); // Unnecessary; condition is always falsy.
+
+const neverNull = {};
+assert(neverNull); // Unnecessary; condition is always truthy.
+```
+
+Whether this option makes sense for your project may vary.
+Some projects may intentionally use assertion functions to ensure that runtime values do indeed match the types according to TypeScript, especially in test code.
+Often, it makes sense to use eslint-disable comments in these cases, with a comment indicating why the condition should be checked at runtime, despite appearing unnecessary.
+However, in some contexts, it may be more appropriate to keep this option disabled entirely.
+
 ### `allowRuleToRunWithoutStrictNullChecksIKnowWhatIAmDoing`
 
 :::danger Deprecated
